StepsTable Component is fully working
